authentication/requirements.txt
django == 4.0.1
asgiref == 3.2.8
djangorestframework ==3.12.4
python-decouple ==3.5
pytz ==2021.3
sqlparse == 0.4.2
django-phonenumbers==1.0.1 
phonenumbers==8.12.54
django-phonenumber==field==6.4.0
djangorestframework==simplejwt-5.2.0
django-templated-mail==1.1.1
djangorestframework==simplejwt-4.8.0
djoser==2.1.0
oauthlib==3.2.0
python3-openid==3.2.0
equests-oauthlib==1.3.1
social-auth-app==django-4.0.0
social-auth-core==4.3.0
